Added method HandleReturnTable like HandleReturn this method automatically sets the error result from BAPIRET or BAPIRET2 but for a return table instead of a structure.
